LWC: Replace Number with Double class

+++ app/src/jvmMain/kotlin/com/saurabhsandav/core/ui/stats/studies/PNLByDayChartStudy.kt
@@ -57,7 +57,7 @@ internal class PNLByDayChartStudy(
                                 month = localDate.monthNumber,
                                 day = localDate.dayOfMonth,
                             ),
-                            value = bigDecimal,
+                            value = bigDecimal.toDouble(),
                         )
                     }
             }
